find_rfp/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.2
dnspython==2.4.1
Faker==19.2.0
numpy==1.24.4
pandas==2.0.3
paramiko==3.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
typing-extensions==4.7.1
tzdata==2023.3
